Split performance tests into TOTP and Token suites
